microvm: virtio-PCI, the modern interface, because vmd has no other
The PCI backend written last commit was legacy 0.9, and read a queue
size of zero from vmd. That was the right answer to the wrong question:
vmd's devices carry PCI ids 0x1040 + type -- 0x1044 for its rng -- and
the specification (virtio 1.3, 4.1.2) is explicit that a
non-transitional device has no legacy interface at all. Only
transitional devices, at 0x1000 + type, answer the 0.9 layout. Legacy
offsets were addressing registers that mean something else.

So this is the modern interface, and almost nothing about it is at a
fixed address. A device describes itself through vendor-specific PCI
capabilities (4.1.4), each naming a BAR and an offset for one
structure: common configuration, notification, ISR, device
configuration. The capability list is walked once at discovery and the
offsets remembered; every access is relative to one of them. Notify is
derived per queue rather than fixed -- cap offset plus the queue's own
notify_off times the capability's multiplier (4.1.4.4).

Three things the seam could not previously express, all now in it:

Features are 64 bits. VERSION_1 is bit 32, so the old uint32_t could
not name the one bit that decides which interface this is. A transport
now declares what it cannot work without -- nothing for legacy mmio,
VERSION_1 for modern PCI -- and negotiation fails honestly if the
device does not offer it, rather than proceeding into a queue setup
that cannot work.

The FEATURES_OK handshake (3.1.1 steps 5 and 6) exists. It is a real
check, not a formality: clearing that bit is the only way a device can
say it cannot live with the subset chosen. Skipped when VERSION_1 was
not negotiated, since a legacy device has no such bit -- and which case
this is falls out of the negotiation rather than being asked
separately.

queue_setup takes the three ring addresses instead of one page number.
That was a legacy concept baked into a seam meant to be transport
neutral; modern names desc, avail and used separately as 64-bit
addresses, and the legacy side recovers its page number from the first
of them, since the three are contiguous either way. The reverse is not
recoverable, which is why the interface is this way round. Modern also
enables a queue as a separate step, which legacy has no equivalent of.

Measured on OpenBSD 8.0-beta: los.platform.rng returns eight bytes, and
four consecutive four-byte reads come back 7f1abae0 00b58933 ef71663b
827d8163 -- distinct, so the queue is cycling rather than handing back
one stale buffer. qemu's mmio path is untouched and its tests still
pass, from the same kernel.

+++ src/platform/microvm/virtio_net.c
@@ -83,7 +83,7 @@ rx_post(int i)
 int
 virtio_net_init(void)
 {
-	uint32_t got = 0;
+	uint64_t got = 0;
 
 	if (virtio_find(VIRTIO_ID_NET, &netdev) != 0)
 		return -1;
